Common things moved to log helpers file
