+++ MANIFEST.in
@@ -435,3 +435,27 @@ recursive-include src *.md
 recursive-include requirements *.txt
 recursive-include src/lightning_app/ui *
 recursive-include src/lightning_app/cli/*-template *
+recursive-include src *.md
+recursive-include requirements *.txt
+recursive-include src/lightning_app/ui *
+recursive-include src/lightning_app/cli/*-template *
+recursive-include src *.md
+recursive-include requirements *.txt
+recursive-include src/lightning_app/ui *
+recursive-include src/lightning_app/cli/*-template *
+recursive-include src *.md
+recursive-include requirements *.txt
+recursive-include src/lightning_app/ui *
+recursive-include src/lightning_app/cli/*-template *
+recursive-include src *.md
+recursive-include requirements *.txt
+recursive-include src/lightning_app/ui *
+recursive-include src/lightning_app/cli/*-template *
+recursive-include src *.md
+recursive-include requirements *.txt
+recursive-include src/lightning_app/ui *
+recursive-include src/lightning_app/cli/*-template *
+recursive-include src *.md
+recursive-include requirements *.txt
+recursive-include src/lightning_app/ui *
+recursive-include src/lightning_app/cli/*-template *
