+++ src/lightning_app/utilities/app_logs.py
@@ -1,15 +1,14 @@
 import json
 import queue
-import sys
 from dataclasses import dataclass
-from datetime import datetime, timedelta
-from json import JSONDecodeError
+from datetime import timedelta
 from threading import Thread
 from typing import Callable, Iterator, List, Optional
 
 import dateutil.parser
 from websocket import WebSocketApp
 
+from lightning_app.utilities.log_helpers import _OrderedLogEntry, _error_callback
 from lightning_app.utilities.logs_socket_api import _LightningLogsSocketAPI
 from lightning_app.utilities.network import LightningClient
 
@@ -27,17 +26,9 @@ class _LogEventLabels:
 
 
 @dataclass
-class _LogEvent:
-    message: str
-    timestamp: datetime
+class _LogEvent(_OrderedLogEntry):
     component_name: str
     labels: _LogEventLabels
-
-    def __ge__(self, other: "_LogEvent") -> bool:
-        return self.timestamp >= other.timestamp
-
-    def __gt__(self, other: "_LogEvent") -> bool:
-        return self.timestamp > other.timestamp
 
 
 def _push_log_events_to_read_queue_callback(component_name: str, read_queue: queue.PriorityQueue):
@@ -63,17 +54,6 @@ def _push_log_events_to_read_queue_callback(component_name: str, read_queue: que
             read_queue.put(event)
 
     return callback
-
-
-def _error_callback(ws_app: WebSocketApp, error: Exception):
-    errors = {
-        KeyError: "Malformed log message, missing key",
-        JSONDecodeError: "Malformed log message",
-        TypeError: "Malformed log format",
-        ValueError: "Malformed date format",
-    }
-    print(f"Error while reading logs ({errors.get(type(error), 'Unknown')})", file=sys.stderr)
-    ws_app.close()
 
 
 def _app_logs_reader(
@@ -130,7 +110,7 @@ def _app_logs_reader(
         pass
 
     except KeyboardInterrupt:
-        # User pressed CTRL+C to exit, we sould respect that
+        # User pressed CTRL+C to exit, we should respect that
         pass
 
     finally:
